pyproject.toml
[tool.poetry]
name = "EPU-LLM"
version = "0.1.0"
description = "Utilzing LLM to construct economic textual indicators."
authors = ["Hsiu-Hsuan Yeh"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]


[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "^1.3.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
orjson = "^3.9.12"
tqdm = "^4.66.1"
pandas = "^2.2.0"
mlflow = "^2.10.2"
openpyxl = "^3.1.2"
langchain = "^0.1.7"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"
ipywidgets = "^8.1.1"
jupyter-bokeh = "^3.0.7"
hvplot = "^0.9.1"
pytest = "^7.4.4"
jupyterlab-execute-time = "2.3.1"
jupyterlab = "3.6.6"
jupyterlab-lsp = "4.3.0"
python-lsp-server = {extras = ["all"], version = "^1.8.2"}
poetry-kernel = "^0.1.3"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
